Update README.md
Line 82, line 104 with questions.

Final goal: compare topics for each posting, save it to the original dataframe so that i can get the occupation category of the posting and other details like no. of applications, and then compare the topics for the postings in each occupation category. 

e.g. For sales persons, are there postings that appear to cover different topics? Are they more or less popular?